+++ packages/components/forma-36-react-datepicker/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1-alpha.58](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.3.1-alpha.57...@contentful/forma-36-react-datepicker@0.3.1-alpha.58) (2020-10-07)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
 ## [0.3.1-alpha.57](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.3.1-alpha.56...@contentful/forma-36-react-datepicker@0.3.1-alpha.57) (2020-10-06)
 
 **Note:** Version bump only for package @contentful/forma-36-react-datepicker